feat: move MagicIntentBundle into core
