+++ magicblock-core/Cargo.toml
@@ -17,7 +17,7 @@ bytes = { workspace = true }
 
 serde = { workspace = true, features = ["derive"] }
 solana-account = { workspace = true }
-solana-account-decoder = { workspace = true }
+solana-account-decoder = { workspace = true, features = ["agave-unstable-api"] }
 solana-clock = { workspace = true }
 solana-hash = { workspace = true }
 solana-message = { workspace = true }
@@ -27,7 +27,7 @@ solana-signature = { workspace = true }
 solana-transaction = { workspace = true, features = ["blake3", "verify"] }
 solana-transaction-context = { workspace = true }
 solana-transaction-error = { workspace = true }
-solana-transaction-status-client-types = { workspace = true }
+solana-transaction-status-client-types = { workspace = true, features = ["agave-unstable-api"] }
 magicblock-magic-program-api = { workspace = true }
 spl-token = { workspace = true }
 spl-token-2022 = { workspace = true }
