+++ programs/magicblock/src/magic_scheduled_base_intent.rs
@@ -1,9 +1,10 @@
 use std::collections::{HashMap, HashSet};
 
 pub use magicblock_core::intent::schedule::{
-    BaseAction, CommitAndUndelegate, CommitType, MagicBaseIntent,
-    MagicIntentBundle, ProgramArgs, UndelegateType, ACTUAL_COMMIT_LIMIT,
-    COMMIT_FEE_LAMPORTS, COMPUTE_UNIT_PRICE_MICRO_LAMPORTS,
+    calculate_commit_fee, BaseAction, CommitAndUndelegate, CommitType,
+    MagicBaseIntent, MagicIntentBundle, ProgramArgs, UndelegateType,
+    ACTUAL_COMMIT_LIMIT, COMMIT_FEE_LAMPORTS,
+    COMPUTE_UNIT_PRICE_MICRO_LAMPORTS,
 };
 use magicblock_core::{
     intent::CommittedAccount,
@@ -13,7 +14,7 @@ use magicblock_core::{
     Slot,
 };
 use magicblock_magic_program_api::args::{
-    ActionArgs, BaseActionArgs, CommitAndUndelegateArgs, CommitTypeArgs,
+    BaseActionArgs, CommitAndUndelegateArgs, CommitTypeArgs,
     MagicBaseIntentArgs, MagicIntentBundleArgs, UndelegateTypeArgs,
 };
 use serde::{Deserialize, Serialize};
@@ -261,10 +262,10 @@ fn validate_magic_intent_bundle_args(
     args: &MagicIntentBundleArgs,
     context: &ConstructionContext<'_, '_, '_>,
 ) -> Result<(), InstructionError> {
-    let committed_set: Option<HashSet<_>> =
-        args.commit.as_ref().map(|el| {
-            el.committed_accounts_indices().iter().copied().collect()
-        });
+    let committed_set: Option<HashSet<_>> = args
+        .commit
+        .as_ref()
+        .map(|el| el.committed_accounts_indices().iter().copied().collect());
     let Some(committed_set) = committed_set else {
         return Ok(());
     };
@@ -566,7 +567,10 @@ impl TryFromArgs<CommitTypeArgs> for CommitType {
                     &accounts,
                     context.transaction_context(),
                 )?;
-                validate_commit_type_accounts(&committed_accounts_ref, context)?;
+                validate_commit_type_accounts(
+                    &committed_accounts_ref,
+                    context,
+                )?;
                 let committed_accounts = committed_accounts_ref
                     .into_iter()
                     .map(|(pubkey, account)| {
@@ -589,7 +593,10 @@ impl TryFromArgs<CommitTypeArgs> for CommitType {
                     &committed_accounts,
                     context.transaction_context(),
                 )?;
-                validate_commit_type_accounts(&committed_accounts_ref, context)?;
+                validate_commit_type_accounts(
+                    &committed_accounts_ref,
+                    context,
+                )?;
 
                 let base_actions = base_actions
                     .into_iter()
@@ -685,44 +692,15 @@ pub(crate) fn validate_commit_schedule_permissions(
     }
 }
 
-pub(crate) fn calculate_commit_fee(
-    accounts: &[CommittedAccount],
-    commit_nonces: &HashMap<Pubkey, u64>,
-) -> Result<u64, InstructionError> {
-    accounts.iter().try_fold(0u64, |fee, account| {
-        if let Some(nonce) = commit_nonces.get(&account.pubkey) {
-            if nonce >= &ACTUAL_COMMIT_LIMIT {
-                Ok(fee + COMMIT_FEE_LAMPORTS)
-            } else {
-                Ok(fee)
-            }
-        } else {
-            Err(InstructionError::Custom(MISSING_COMMIT_NONCE_ERR))
-        }
-    })
-}
-
-fn calculate_actions_fee(actions: &[BaseAction]) -> u64 {
-    const MICRO_LAMPORTS_PER_LAMPORT: u64 = 1_000_000;
-    let micro_lamports = actions.iter().fold(0u64, |acc, action| {
-        acc.saturating_add(
-            action.compute_units as u64 * COMPUTE_UNIT_PRICE_MICRO_LAMPORTS,
-        )
-    });
-    micro_lamports.div_ceil(MICRO_LAMPORTS_PER_LAMPORT)
-}
-
 #[cfg(test)]
 mod tests {
     use std::collections::HashMap;
 
     use magicblock_core::intent::CommittedAccount;
     use solana_account::Account;
-    use solana_instruction::error::InstructionError;
     use solana_pubkey::Pubkey;
 
     use super::*;
-    use crate::magic_sys::MISSING_COMMIT_NONCE_ERR;
 
     fn make_committed_account(pubkey: Pubkey) -> CommittedAccount {
         CommittedAccount {
@@ -745,77 +723,6 @@ mod tests {
             account_metas_per_program: vec![],
             callback: None,
         }
-    }
-
-    // ---- calculate_commit_fee ----
-
-    #[test]
-    fn test_commit_fee_at_limit_is_zero() {
-        let pk = Pubkey::new_unique();
-        // nonce is commits done so far; nonce+1 is the next commit number.
-        // ACTUAL_COMMIT_LIMIT - 1 means the next commit is exactly at the limit → free.
-        let nonces = HashMap::from([(pk, ACTUAL_COMMIT_LIMIT - 1)]);
-        let fee = calculate_commit_fee(&[make_committed_account(pk)], &nonces)
-            .unwrap();
-        assert_eq!(fee, 0);
-    }
-
-    #[test]
-    fn test_commit_fee_above_limit_charges_per_account() {
-        let pk1 = Pubkey::new_unique();
-        let pk2 = Pubkey::new_unique();
-        let nonces = HashMap::from([
-            (pk1, ACTUAL_COMMIT_LIMIT + 1),
-            (pk2, ACTUAL_COMMIT_LIMIT + 1),
-        ]);
-        let fee = calculate_commit_fee(
-            &[make_committed_account(pk1), make_committed_account(pk2)],
-            &nonces,
-        )
-        .unwrap();
-        assert_eq!(fee, COMMIT_FEE_LAMPORTS * 2);
-    }
-
-    #[test]
-    fn test_commit_fee_mixed_accounts() {
-        let pk_below = Pubkey::new_unique();
-        let pk_above = Pubkey::new_unique();
-        let nonces = HashMap::from([
-            (pk_below, ACTUAL_COMMIT_LIMIT - 1), // next commit is exactly at limit → free
-            (pk_above, ACTUAL_COMMIT_LIMIT), // next commit exceeds limit → charged
-        ]);
-        let fee = calculate_commit_fee(
-            &[
-                make_committed_account(pk_below),
-                make_committed_account(pk_above),
-            ],
-            &nonces,
-        )
-        .unwrap();
-        assert_eq!(fee, COMMIT_FEE_LAMPORTS);
-    }
-
-    #[test]
-    fn test_commit_fee_missing_nonce_errors() {
-        let pk = Pubkey::new_unique();
-        let err = calculate_commit_fee(
-            &[make_committed_account(pk)],
-            &HashMap::new(),
-        )
-        .unwrap_err();
-        assert_eq!(err, InstructionError::Custom(MISSING_COMMIT_NONCE_ERR));
-    }
-
-    /// two actions of 200_000 CUs each → 20_000 lamports
-    #[test]
-    fn test_actions_fee_multiple_actions() {
-        assert_eq!(
-            calculate_actions_fee(&[
-                make_base_action(200_000),
-                make_base_action(200_000)
-            ]),
-            20_000
-        );
     }
 
     // ---- ScheduledIntentBundle::calculate_fee ----
